refactor: feature内の配置をポリシーに合わせる
component indexの公開範囲を同名コンポーネントに絞り、MonthlyTotals固有の取得処理をslice配下へ移動した。
