+++ apps/web/src/features/budgets/components/MonthlyBudgetUsage/index.ts
@@ -1,2 +1,1 @@
 export { MonthlyBudgetUsage } from "./MonthlyBudgetUsage"
-export type { MonthlyBudgetUsageProps } from "./MonthlyBudgetUsage"
