+++ apps/web/src/features/categories/components/CategorySelect/index.ts
@@ -1,7 +1,1 @@
-export {
-  CategorySelect,
-  CategoryOption,
-  NoneCategoryOption,
-  ErrorCategoryOption,
-  LoadingCategoryOption,
-} from "./CategorySelect"
+export { CategorySelect } from "./CategorySelect"
