+++ apps/web/src/features/categories/index.ts
@@ -1,9 +1,9 @@
 export {
   CategoryOption,
-  CategorySelect,
   ErrorCategoryOption,
   LoadingCategoryOption,
-} from "./components/CategorySelect"
+} from "./components/CategorySelect/CategorySelect"
+export { CategorySelect } from "./components/CategorySelect"
 export { CategorySettingsList } from "./components/CategorySettingsList"
 export { useCategories } from "./listCategory/useCategories"
 export { categoryQueryKeys } from "./queryKeys"
